Refactor tests by moving shared setup to utils.py
